Выносит кнопку в отдельный компонент
Добавляет текущий баланс в шашборд
