+++ src/components/AddTransactionBtn/Style.module.css
@@ -1,0 +1,5 @@
+.addBtnBg {
+  position: absolute;
+  bottom: 30px;
+  right: 30px;
+}
