+++ src/components/AddTransactionBtn/img/index.js
@@ -1,0 +1,3 @@
+import addBtnBg from './addBtnBg.svg';
+
+export const BtnImg = { addBtnBg };
