+++ src/components/Table/Style.module.css
@@ -26,7 +26,7 @@
     line-height: 1.437;
     border-top-right-radius: 10px;
     border-bottom-color: rgb(189, 189, 189);
-    /* display: block; */
+
     text-align: right;
   }
 
@@ -39,5 +39,9 @@
     /* content: none; */
 
     float: left;
+  }
+
+  td {
+    background-color: #fff;
   }
 }
