+++ src/pages/DashboardPage/DashboardPage.js
@@ -1,44 +1,32 @@
 import React from 'react';
-import { Button, Container } from 'react-bootstrap';
+import { Button } from 'react-bootstrap';
 import {
   Header,
   Navigation,
   Currency,
+  Balance,
   OperationsTable,
   ContainerAppWrapper,
+  AddTransactionBtn,
 } from '../../components';
 import { dashBoardImg } from './img';
 
 import style from './Style.module.css';
 
-const inlineStyles = {
-  addBtn: {
-    backgroundColor: '#24CCA7',
-    width: '44px',
-    height: '44px',
-    borderRadius: '50%',
-    padding: '0px',
-    border: 'none',
-  },
-};
-
 const DashboardPage = () => {
   return (
-    <>
-      <ContainerAppWrapper>
-        <Header />
-        <div className={style.dashboardPage}>
-          <Navigation />
-          <Currency />
-          <OperationsTable />
-          <img className={style.elipse_top} src={dashBoardImg.ellipseTop} alt="" />
-          <img className={style.elipse_bottom} src={dashBoardImg.ellipseBottom} alt="" />
-        </div>
-      </ContainerAppWrapper>
-      <Button className={style.addBtnBg} style={inlineStyles.addBtn}>
-        <img src={dashBoardImg.addBtnBg} alt="" />
-      </Button>
-    </>
+    <ContainerAppWrapper>
+      <Header />
+      <div className={style.dashboardPage}>
+        <Navigation />
+        <Currency />
+        <Balance />
+        <OperationsTable />
+        <img className={style.elipse_top} src={dashBoardImg.ellipseTop} alt="" />
+        <img className={style.elipse_bottom} src={dashBoardImg.ellipseBottom} alt="" />
+        <AddTransactionBtn />
+      </div>
+    </ContainerAppWrapper>
   );
 };
