+++ src/pages/DashboardPage/Style.module.css
@@ -1,18 +1,4 @@
-.addBtnBg {
-  /* position: absolute;
-  width: 44px;
-  height: 44px;
-  right: 40px;
-  bottom: 40px; */
-  position: fixed;
-  top: 80%;
-  left: 240px;
-}
-
 .dashboardPage {
-  /* display: flex;
-  justify-content: space-between; */
-
   background-repeat: no-repeat;
   background-color: rgba(255, 255, 255, 0.4);
   backdrop-filter: blur(50px);
@@ -22,8 +8,6 @@
 }
 @media (max-width: 768px) {
   .dashboardPage {
-    /* display: flex;
-  justify-content: space-between; */
     min-height: 720px;
     background-color: rgba(255, 255, 255, 0.4);
     backdrop-filter: blur(50px);
