+++ src/pages/DashboardPage/img/index.js
@@ -1,5 +1,4 @@
-import addBtnBg from './addBtnBg.svg';
 import ellipseTop from './ellipseTop.svg';
 import ellipseBottom from './ellipseBottom.svg';
 
-export const dashBoardImg = { addBtnBg, ellipseTop, ellipseBottom };
+export const dashBoardImg = { ellipseTop, ellipseBottom };
